Added breakpoints to theme object and refactored declaration file to fix theme object not maintaining types when writing css
