+++ src/theme/index.ts
@@ -1,8 +1,11 @@
+import { breakpoints } from './breakpoints'
+
 const theme = {
 	colors: {
 		black: '#000',
 		white: '#fff',
 	},
+	breakpoints,
 } as const
 
 export default theme
